Extract output path rules
